# swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Swagger CocktailsStore - OpenAPI 3.0",
		"description": "This is a simple Cocktails Store Server based on the OpenAPI 3.0 specification.  You can find out more about\nSwagger at [https://swagger.io](https://swagger.io). In the third iteration of the cocktails store, we've switched to the design first approach!\nYou can now help us improve the API whether it's by making changes to the definition itself or to the code.\nThat way, with time, we can improve the API in general, and expose some of the new features in OAS3.\n\n\nSome useful links:\n- [Cocktails Store repository](https://github.com/MagisterUnivers/TeamName-project)",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"email": "apiteam@swagger.io"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		},
		"version": "1.0.11"
	},
...
	"servers": [
		{
			"url": "https://cocktails-backend-cwrh.onrender.com"
		},
		{ "url": "http://localhost:3001" }
	],
...
	"paths": {
...
		"/users/theme": {
			"patch": {
				"tags": ["User"],
				"summary": "Update theme update",
				"description": "This can be done only by logged in user",
				"security": [{ "Bearer": [] }],
				"operationId": "updateTheme",
				"parameters": [
					{
						"name": "token",
						"in": "header",
						"description": "Token from database. ",
						"required": true,
						"schema": {
							"type": "string",
							"example": "token123cnadjnw"
						}
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"example": {
										"theme": "theme"
									}
								}
							},
							"application/xml": {
								"schema": {
									"type": "string",
									"example": {
										"theme": "theme"
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid token supplied"
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "User not found"
					}
				}
			}
		},
...
	},
	"components": {
...
		"securitySchemes": {
			"Bearer": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}